backend/requirements.txt
alembic==1.16.5
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.2
PyMySQL==1.1.2
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
rpds-py==0.27.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
